+++ phidgets_api/CMakeLists.txt
@@ -1,30 +1,20 @@
-cmake_minimum_required(VERSION 2.8.3)
+cmake_minimum_required(VERSION 3.5)
 
 project(phidgets_api)
 
-# Use C++14
-if (CMAKE_VERSION VERSION_LESS "3.1")
-    if (CMAKE_CXX_COMPILER_ID STREQUAL "GNU")
-        set(CMAKE_CXX_FLAGS "-std=gnu++14 ${CMAKE_CXX_FLAGS}")
-    endif()
-else()
-    set (CMAKE_CXX_STANDARD 14)
+# Default to C++14
+if(NOT CMAKE_CXX_STANDARD)
+  set(CMAKE_CXX_STANDARD 14)
 endif()
 
-# High level of warnings:
-if (CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
-    add_compile_options(-Wall -Wextra)
+if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
+  add_compile_options(-Wall -Wextra -Wpedantic)
 endif()
 
-find_package(catkin REQUIRED COMPONENTS libphidget22)
+find_package(ament_cmake_ros REQUIRED)
+find_package(libphidget22 REQUIRED)
 
-catkin_package(
-    INCLUDE_DIRS include
-    LIBRARIES phidgets_api
-    CATKIN_DEPENDS libphidget22
-)
-
-include_directories(include ${catkin_INCLUDE_DIRS})
+include_directories(include)
 
 add_library(phidgets_api src/accelerometer.cpp
                          src/analog_input.cpp
@@ -44,17 +34,22 @@ add_library(phidgets_api src/accelerometer.cpp
                          src/spatial.cpp
                          src/temperature.cpp)
 
-add_dependencies(phidgets_api ${catkin_EXPORTED_TARGETS})
-
-target_link_libraries(phidgets_api ${catkin_LIBRARIES})
+ament_target_dependencies(phidgets_api
+  libphidget22
+)
 
 install(TARGETS phidgets_api
-    ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
-    LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
-    RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
+    ARCHIVE DESTINATION lib
+    LIBRARY DESTINATION lib
+    RUNTIME DESTINATION bin
 )
 
-install(DIRECTORY include/${PROJECT_NAME}/
-    DESTINATION ${CATKIN_PACKAGE_INCLUDE_DESTINATION}
-    FILES_MATCHING PATTERN "*.h"
+install(DIRECTORY include/
+    DESTINATION include
 )
+
+ament_export_dependencies(ament_cmake libphidget22)
+ament_export_include_directories(include)
+ament_export_libraries(phidgets_api)
+
+ament_package()
