+++ phidgets_api/include/phidgets_api/digital_input.hpp
@@ -27,51 +27,41 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef PHIDGETS_API_TEMPERATURE_H
-#define PHIDGETS_API_TEMPERATURE_H
+#ifndef PHIDGETS_API_DIGITAL_INPUT_H
+#define PHIDGETS_API_DIGITAL_INPUT_H
 
 #include <functional>
 
 #include <libphidget22/phidget22.h>
 
-#include "phidgets_api/phidget22.h"
+#include "phidgets_api/phidget22.hpp"
 
 namespace phidgets {
 
-enum class ThermocoupleType {
-    J_TYPE = 1,
-    K_TYPE = 2,
-    E_TYPE = 3,
-    T_TYPE = 4,
-};
-
-class Temperature final
+class DigitalInput
 {
   public:
-    PHIDGET22_NO_COPY_NO_MOVE_NO_ASSIGN(Temperature)
+    PHIDGET22_NO_COPY_NO_MOVE_NO_ASSIGN(DigitalInput)
 
-    explicit Temperature(int32_t serial_number, int hub_port,
-                         bool is_hub_port_device,
-                         std::function<void(double)> temperature_handler);
+    explicit DigitalInput(int32_t serial_number, int hub_port,
+                          bool is_hub_port_device, int channel,
+                          std::function<void(int, int)> input_handler);
 
-    ~Temperature();
+    ~DigitalInput();
 
-    void setThermocoupleType(ThermocoupleType type);
+    bool getInputValue() const;
 
-    double getTemperature() const;
-
-    void setDataInterval(uint32_t interval_ms) const;
-
-    void temperatureChangeHandler(double temperature) const;
+    void stateChangeHandler(int state) const;
 
   private:
-    std::function<void(double)> temperature_handler_;
-    PhidgetTemperatureSensorHandle temperature_handle_;
-    static void TemperatureChangeHandler(
-        PhidgetTemperatureSensorHandle temperature_handle, void *ctx,
-        double temperature);
+    int channel_;
+    std::function<void(int, int)> input_handler_;
+    PhidgetDigitalInputHandle di_handle_;
+
+    static void StateChangeHandler(PhidgetDigitalInputHandle input_handle,
+                                   void *ctx, int state);
 };
 
 }  // namespace phidgets
 
-#endif  // PHIDGETS_API_TEMPERATURE_H
+#endif  // PHIDGETS_API_DIGITAL_INPUT_H
