+++ phidgets_api/include/phidgets_api/encoder.hpp
@@ -34,7 +34,7 @@
 
 #include <libphidget22/phidget22.h>
 
-#include "phidgets_api/phidget22.h"
+#include "phidgets_api/phidget22.hpp"
 
 namespace phidgets {
