+++ phidgets_api/include/phidgets_api/magnetometer.hpp
@@ -27,37 +27,51 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef PHIDGETS_API_IR_H
-#define PHIDGETS_API_IR_H
+#ifndef PHIDGETS_API_MAGNETOMETER_H
+#define PHIDGETS_API_MAGNETOMETER_H
 
 #include <functional>
 
 #include <libphidget22/phidget22.h>
 
-#include "phidgets_api/phidget22.h"
+#include "phidgets_api/phidget22.hpp"
 
 namespace phidgets {
 
-class IR final
+class Magnetometer final
 {
   public:
-    PHIDGET22_NO_COPY_NO_MOVE_NO_ASSIGN(IR)
+    PHIDGET22_NO_COPY_NO_MOVE_NO_ASSIGN(Magnetometer)
 
-    explicit IR(int32_t serial_number,
-                std::function<void(const char *, uint32_t, int)> code_handler);
+    explicit Magnetometer(
+        int32_t serial_number, int hub_port, bool is_hub_port_device,
+        std::function<void(const double[3], double)> data_handler);
 
-    ~IR();
+    ~Magnetometer();
 
-    void codeHandler(const char *code, uint32_t bit_count, int is_repeat) const;
+    void setCompassCorrectionParameters(double cc_mag_field, double cc_offset0,
+                                        double cc_offset1, double cc_offset2,
+                                        double cc_gain0, double cc_gain1,
+                                        double cc_gain2, double cc_T0,
+                                        double cc_T1, double cc_T2,
+                                        double cc_T3, double cc_T4,
+                                        double cc_T5);
+
+    void getMagneticField(double &x, double &y, double &z,
+                          double &timestamp) const;
+
+    void setDataInterval(uint32_t interval_ms) const;
+
+    void dataHandler(const double magnetic_field[3], double timestamp) const;
 
   private:
-    std::function<void(const char *, uint32_t, int)> code_handler_;
-    PhidgetIRHandle ir_handle_;
+    std::function<void(const double[3], double)> data_handler_;
+    PhidgetMagnetometerHandle mag_handle_;
 
-    static void CodeHandler(PhidgetIRHandle ir, void *ctx, const char *code,
-                            uint32_t bit_count, int is_repeat);
+    static void DataHandler(PhidgetMagnetometerHandle input_handle, void *ctx,
+                            const double magnetic_field[3], double timestamp);
 };
 
 }  // namespace phidgets
 
-#endif  // PHIDGETS_API_IR_H
+#endif  // PHIDGETS_API_MAGNETOMETER_H
