+++ phidgets_api/include/phidgets_api/temperature.hpp
@@ -27,31 +27,51 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef PHIDGETS_API_DIGITAL_OUTPUT_H
-#define PHIDGETS_API_DIGITAL_OUTPUT_H
+#ifndef PHIDGETS_API_TEMPERATURE_H
+#define PHIDGETS_API_TEMPERATURE_H
+
+#include <functional>
 
 #include <libphidget22/phidget22.h>
 
-#include "phidgets_api/phidget22.h"
+#include "phidgets_api/phidget22.hpp"
 
 namespace phidgets {
 
-class DigitalOutput final
+enum class ThermocoupleType {
+    J_TYPE = 1,
+    K_TYPE = 2,
+    E_TYPE = 3,
+    T_TYPE = 4,
+};
+
+class Temperature final
 {
   public:
-    PHIDGET22_NO_COPY_NO_MOVE_NO_ASSIGN(DigitalOutput)
+    PHIDGET22_NO_COPY_NO_MOVE_NO_ASSIGN(Temperature)
 
-    explicit DigitalOutput(int32_t serial_number, int hub_port,
-                           bool is_hub_port_device, int channel);
+    explicit Temperature(int32_t serial_number, int hub_port,
+                         bool is_hub_port_device,
+                         std::function<void(double)> temperature_handler);
 
-    ~DigitalOutput();
+    ~Temperature();
 
-    void setOutputState(bool state) const;
+    void setThermocoupleType(ThermocoupleType type);
+
+    double getTemperature() const;
+
+    void setDataInterval(uint32_t interval_ms) const;
+
+    void temperatureChangeHandler(double temperature) const;
 
   private:
-    PhidgetDigitalOutputHandle do_handle_;
+    std::function<void(double)> temperature_handler_;
+    PhidgetTemperatureSensorHandle temperature_handle_;
+    static void TemperatureChangeHandler(
+        PhidgetTemperatureSensorHandle temperature_handle, void *ctx,
+        double temperature);
 };
 
 }  // namespace phidgets
 
-#endif  // PHIDGETS_API_DIGITAL_OUTPUT_H
+#endif  // PHIDGETS_API_TEMPERATURE_H
