+++ phidgets_api/src/analog_input.cpp
@@ -34,8 +34,8 @@
 
 #include <libphidget22/phidget22.h>
 
-#include "phidgets_api/analog_input.h"
-#include "phidgets_api/phidget22.h"
+#include "phidgets_api/analog_input.hpp"
+#include "phidgets_api/phidget22.hpp"
 
 namespace phidgets {
