+++ phidgets_api/src/analog_inputs.cpp
@@ -34,9 +34,9 @@
 
 #include <libphidget22/phidget22.h>
 
-#include "phidgets_api/analog_input.h"
-#include "phidgets_api/analog_inputs.h"
-#include "phidgets_api/phidget22.h"
+#include "phidgets_api/analog_input.hpp"
+#include "phidgets_api/analog_inputs.hpp"
+#include "phidgets_api/phidget22.hpp"
 
 namespace phidgets {
