+++ phidgets_api/src/encoder.cpp
@@ -31,8 +31,8 @@
 
 #include <libphidget22/phidget22.h>
 
-#include "phidgets_api/encoder.h"
-#include "phidgets_api/phidget22.h"
+#include "phidgets_api/encoder.hpp"
+#include "phidgets_api/phidget22.hpp"
 
 namespace phidgets {
