+++ phidgets_api/src/encoders.cpp
@@ -31,9 +31,9 @@
 #include <memory>
 #include <vector>
 
-#include "phidgets_api/encoder.h"
-#include "phidgets_api/encoders.h"
-#include "phidgets_api/phidget22.h"
+#include "phidgets_api/encoder.hpp"
+#include "phidgets_api/encoders.hpp"
+#include "phidgets_api/phidget22.hpp"
 
 namespace phidgets {
