+++ phidgets_api/src/motor.cpp
@@ -32,8 +32,8 @@
 
 #include <libphidget22/phidget22.h>
 
-#include "phidgets_api/motor.h"
-#include "phidgets_api/phidget22.h"
+#include "phidgets_api/motor.hpp"
+#include "phidgets_api/phidget22.hpp"
 
 namespace phidgets {
