+++ phidgets_api/src/motors.cpp
@@ -32,8 +32,8 @@
 
 #include <libphidget22/phidget22.h>
 
-#include "phidgets_api/motors.h"
-#include "phidgets_api/phidget22.h"
+#include "phidgets_api/motors.hpp"
+#include "phidgets_api/phidget22.hpp"
 
 namespace phidgets {
