+++ phidgets_api/src/phidget22.cpp
@@ -29,7 +29,7 @@
 
 #include <string>
 
-#include "phidgets_api/phidget22.h"
+#include "phidgets_api/phidget22.hpp"
 
 namespace phidgets {
